+++ bin/miden-cli/src/commands/new_transactions.rs
@@ -1,4 +1,4 @@
-use std::io;
+use std::{io, sync::Arc};
 
 use clap::{Parser, ValueEnum};
 use miden_client::{
@@ -12,13 +12,14 @@ use miden_client::{
     },
     Client,
 };
+use miden_tx_prover::RemoteTransactionProver;
 use tracing::info;
 
 use crate::{
     create_dynamic_table,
     utils::{
-        get_input_acc_id_by_prefix_or_default, load_faucet_details_map, parse_account_id,
-        SHARED_TOKEN_DOCUMENTATION,
+        get_input_acc_id_by_prefix_or_default, load_config_file, load_faucet_details_map,
+        parse_account_id, SHARED_TOKEN_DOCUMENTATION,
     },
 };
 
@@ -53,6 +54,10 @@ pub struct MintCmd {
     /// Flag to submit the executed transaction without asking for confirmation
     #[clap(long, default_value_t = false)]
     force: bool,
+
+    /// Flag to delegate proving to the remote prover specified in the config file
+    #[clap(long, default_value_t = false)]
+    delegate_proving: bool,
 }
 
 impl MintCmd {
@@ -73,8 +78,14 @@ impl MintCmd {
         .map_err(|err| err.to_string())?
         .build();
 
-        execute_transaction(&mut client, fungible_asset.faucet_id(), transaction_request, force)
-            .await
+        execute_transaction(
+            &mut client,
+            fungible_asset.faucet_id(),
+            transaction_request,
+            force,
+            self.delegate_proving,
+        )
+        .await
     }
 }
 
@@ -104,6 +115,10 @@ pub struct SendCmd {
     /// Setting this flag turns the transaction from a PayToId to a PayToIdWithRecall.
     #[clap(short, long)]
     recall_height: Option<u32>,
+
+    /// Flag to delegate proving to the remote prover specified in the config file
+    #[clap(long, default_value_t = false)]
+    delegate_proving: bool,
 }
 
 impl SendCmd {
@@ -134,7 +149,14 @@ impl SendCmd {
         .map_err(|err| err.to_string())?
         .build();
 
-        execute_transaction(&mut client, sender_account_id, transaction_request, force).await
+        execute_transaction(
+            &mut client,
+            sender_account_id,
+            transaction_request,
+            force,
+            self.delegate_proving,
+        )
+        .await
     }
 }
 
@@ -159,6 +181,10 @@ pub struct SwapCmd {
     /// Flag to submit the executed transaction without asking for confirmation
     #[clap(long, default_value_t = false)]
     force: bool,
+
+    /// Flag to delegate proving to the remote prover specified in the config file
+    #[clap(long, default_value_t = false)]
+    delegate_proving: bool,
 }
 
 impl SwapCmd {
@@ -190,7 +216,14 @@ impl SwapCmd {
         .map_err(|err| err.to_string())?
         .build();
 
-        execute_transaction(&mut client, sender_account_id, transaction_request, force).await?;
+        execute_transaction(
+            &mut client,
+            sender_account_id,
+            transaction_request,
+            force,
+            self.delegate_proving,
+        )
+        .await?;
 
         let payback_note_tag: u32 = build_swap_tag(
             (&self.note_type).into(),
@@ -222,6 +255,10 @@ pub struct ConsumeNotesCmd {
     /// Flag to submit the executed transaction without asking for confirmation
     #[clap(short, long, default_value_t = false)]
     force: bool,
+
+    /// Flag to delegate proving to the remote prover specified in the config file
+    #[clap(long, default_value_t = false)]
+    delegate_proving: bool,
 }
 
 impl ConsumeNotesCmd {
@@ -252,7 +289,14 @@ impl ConsumeNotesCmd {
 
         let transaction_request = TransactionRequestBuilder::consume_notes(list_of_notes).build();
 
-        execute_transaction(&mut client, account_id, transaction_request, force).await
+        execute_transaction(
+            &mut client,
+            account_id,
+            transaction_request,
+            force,
+            self.delegate_proving,
+        )
+        .await
     }
 }
 
@@ -264,6 +308,7 @@ async fn execute_transaction(
     account_id: AccountId,
     transaction_request: TransactionRequest,
     force: bool,
+    delegated_proving: bool,
 ) -> Result<(), String> {
     println!("Executing transaction...");
     let transaction_execution_result =
@@ -291,9 +336,22 @@ async fn execute_transaction(
         .map(|note| note.id())
         .collect::<Vec<_>>();
 
-    client.submit_transaction(transaction_execution_result).await?;
+    if delegated_proving {
+        let (cli_config, _) = load_config_file()?;
+        let remote_prover_endpoint = cli_config
+            .remote_prover_endpoint
+            .as_ref()
+            .ok_or("Remote prover endpoint not found in config file")?;
+        let remote_prover =
+            Arc::new(RemoteTransactionProver::new(&remote_prover_endpoint.to_string()));
+        client
+            .submit_transaction_with_prover(transaction_execution_result, remote_prover)
+            .await?;
+    } else {
+        client.submit_transaction(transaction_execution_result).await?;
+    }
 
-    println!("Succesfully created transaction.");
+    println!("Successfully created transaction.");
     println!("Transaction ID: {}", transaction_id);
 
     if output_notes.is_empty() {