+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 * Added `TransactionRequestBuilder` for building `TransactionRequest` (#605).
 * Added Sync Loop to Integration Tests for Small Speedup (#590).
 
+### Changes
+* [BREAKING] Added per transaction prover support to the client (#599).
+
 ## 0.6.0 (2024-11-08)
 
 ### Features